refactor: move mock-logger.ts from __test__/ to src/test-utils/
