+++ src/cli/commands/generate.test.ts
@@ -7,8 +7,8 @@ import { IgnoreProcessor } from "../../features/ignore/ignore-processor.js";
 import { McpProcessor } from "../../features/mcp/mcp-processor.js";
 import { RulesProcessor } from "../../features/rules/rules-processor.js";
 import { SubagentsProcessor } from "../../features/subagents/subagents-processor.js";
+import { createMockLogger } from "../../test-utils/mock-logger.js";
 import { fileExists } from "../../utils/file.js";
-import { createMockLogger } from "./__test__/mock-logger.js";
 import type { GenerateOptions } from "./generate.js";
 import { generateCommand } from "./generate.js";
